Fix #38: Add the possibility to configure how Pods are scheduled
Problem: Today there's no way of configuring how Pods are scheduled

Solution:

Add configurable nodeSelector and tolerations to the
Deployment template and values file.

I've added empty values as default to the values file,
there's also commented examples on how this can be done. In the
case of the examples they force scheduling of the controller Pod(s)
onto the master nodes in the cluster.

If nodeSelector and tolerations are set in the values file they're
rendered in the final manifest generated with helm.

Testing (optional if not described in Solution section):